+++ src/program-instructions.ts
@@ -88,6 +88,8 @@ export async function depositIx(
       authority: userKey,
       tokenProgram: TOKEN_PROGRAM_ID,
       state: Exchange.stateAddress,
+      greeks: Exchange.zetaGroup.greeks,
+      oracle: Exchange.zetaGroup.oracle,
     },
     remainingAccounts,
   });